Skrrt Skrrt
All of my feelings hurt hurt.

So shits been tidied up and thrown around a bit, I need to make some changes to it still and try and find a way to get the dropdown to work but that's not something I wanna put 200 hours into right now to see if it's gonna work or not.